requirements.txt
anyio==3.6.1
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
configobj==5.0.6
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
humanize==4.3.0
idna==3.3
numpy==1.23.1
pandas==1.4.3
passlib==1.7.4
pgcli==1.6.0
pgspecial==2.0.1
prompt-toolkit==1.0.18
psycopg==3.0.16
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
Pygments==2.13.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
starlette==0.19.1
typing==3.7.4.3
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
wcwidth==0.2.5
